+++ RELEASE_NOTES.md
@@ -3,6 +3,7 @@
 Changes:
 
   * Allowed assigning a barcode to loose storage
+  * Improved performance of Edit Project page
   * Fixed RNA Library Preparation spreadsheet download and Pinery-MISO excluding some QCs if QC type
     name doesn't exactly match the original spec.
   * Fixed error displaying/validating QC type when editing QC with an archived QC type